+++ mz_bokeh_package/custom_widgets/__init__.py
@@ -1,0 +1,2 @@
+from .custom_select import CustomSelect  # noqa F401
+from .custom_multiselect import CustomMultiSelect  # noqa F401
